+++ package.js
@@ -32,7 +32,6 @@ function configure( api ){
     _use( 'blaze-html-templates@2.0.0 || 3.0.0-alpha300.0', 'client' );
     _use( 'ecmascript' );
     _use( 'less@4.0.0', 'client' );
-    _use( 'pwix:orderable-stack@1.0.1' );
     _use( 'tmeasday:check-npm-versions@1.0.2 || 2.0.0-beta.0', 'server' );
     api.addFiles( 'src/client/components/TypedMessage/TypedMessage.js', 'client' );
 }