+++ src/common/classes/base.class.js
@@ -1,0 +1,26 @@
+/*
+ * pwix:typed-message/src/common/classes/base.class.js
+ *
+ * A base class which does nothing but because '@vestergaard-company/js-mixin' only lets add an interface when extending a class...
+ */
+
+export class Base {
+
+    // static data
+
+    // static methods
+
+    // private data
+
+    // private methods
+
+    // public data
+
+    /**
+     * Constructor
+     * @returns {Base} this instance
+     */
+    constructor(){
+        return this;
+    }
+}