+++ src/common/classes/typed-message.class.js
@@ -1,20 +1,18 @@
 /*
  * pwix:typed-message/src/common/classes/typed-message.class.js
  *
- * A class which implements the ITypedMessage interface.
+ * A simple class which implements the ITypedMessage interface.
  */
 
 import _ from 'lodash';
 const assert = require( 'assert' ).strict; // up to nodejs v16.x
 import mix from '@vestergaard-company/js-mixin';
 
-import { OStack } from 'meteor/pwix:orderable-stack';
-
 import { ITypedMessage } from '../interfaces/ityped-message.iface.js';
 
-import { LevelOrder } from '../definitions/level-order.def.js';
+import { Base } from './base.class.js';
 
-export class TypedMessage  extends  mix( OStack.Orderable ).with( ITypedMessage ){
+export class TypedMessage  extends  mix( Base ).with( ITypedMessage ){
 
     // static data
 
@@ -28,21 +26,10 @@ export class TypedMessage  extends  mix( OStack.Orderable ).with( ITypedMessage 
 
     /**
      * Constructor
-     * @returns {TypedMessage}
+     * @returns {TypedMessage} this instance
      */
     constructor(){
         super( ...arguments );
         return this;
-    }
-
-    // the method each IOrderable implementation must provide
-    // here we keep the usual semantic where 'EMERGENCY' is greater (has a higher priority level) than 'DEBUG'
-    //  so sort the levels in the reverse order of their index
-    iOrderableCompare( a, b ){
-        assert( a instanceof ITypedMessage, 'expected an ITypedMessage instance' );
-        assert( b instanceof ITypedMessage, 'expected an ITypedMessage instance' );
-        const level_a = a.ITypedMessageType();
-        const level_b = b.ITypedMessageType();
-        return -1 * LevelOrder.compare( level_a, level_b );
     }
 }